src/requirements.txt
fastapi==0.125.0
uvicorn==0.38.0
python-multipart==0.0.21
python-dotenv==1.2.1
pydantic-settings==2.12.0
aiofiles==25.1.0
langchain==1.2.0
langchain-community==0.4.1
langchain-text-splitters==1.1.0
PyMuPDF==1.26.7
docx2txt==0.8
openai==2.14.0
cohere==5.20.1
groq==1.0.0
qdrant-client==1.16.2
SQLAlchemy==2.0.45
asyncpg==0.31.0
alembic==1.18.0
psycopg2-binary==2.9.11
pgvector==0.4.2
nltk==3.9.2
pymongo==4.16.0

# Monitoring and metrics
prometheus-client==0.24.1
starlette-exporter==0.23.0
fastapi-health==0.4.0

# Task Queue and Background Processing
celery==5.6.2
redis==7.1.0
kombu==5.6.2
billiard==4.2.4
vine==5.1.0
flower==2.0.1

# Markdown and Web URL ingestion
Markdown==3.7
beautifulsoup4==4.13.4
lxml==5.4.0
httpx==0.28.1